agents/pyproject.toml
[project]
name = "langgraph-research-agent"
version = "0.1.0"
description = "A Deep Reseach Agent built with Langgraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=13.19.2",
    "arize-phoenix-client>=2.1.0",
    "arize-phoenix-otel>=0.15.0",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.11.4",
    "deepagents>=0.4.12",
    "elasticsearch>=9.3.0",
    "langchain>=1.2.13",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "markdownify>=1.2.2",
    "networkx>=3.6.1",
    "notebook>=7.5.5",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "openinference-instrumentation-langchain>=0.1.61",
    "openinference-instrumentation-openai>=0.1.42",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "sqlalchemy>=2.0.48",
    "tavily-python>=0.7.23",
    "typing-extensions>=4.15.0",
    "weaviate-client>=4.20.4",
    "wikipedia-api>=0.11.0",
]
